move test utility functions in common folder
